requirements.txt
streamlit>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
ultralytics>=8.0.0